core side for [DP6243]
git-svn-id: http://svn.l2jserver.com/trunk@3155 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/net/sf/l2j/gameserver/model/L2TradeList.java
@@ -68,7 +68,7 @@ public class L2TradeList
         }
 	}
 
-    public void replaceItem(int itemID, int price)
+    public void replaceItem(int itemID, long price)
     {
         L2TradeItem item = _items.get(itemID);
         if (item != null)
